+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - feat(nextjs): Add instructions on how to add a `global-error` page to Next.js
   App Router (#506)
+- feat(nextjs): Automatically enable vercel cron monitors (#507)
 
 ## 3.17.0